feat: fix chromosomes script can handle uncompressed files

+++ data/scripts/fix_chromosomes.sh
@@ -5,12 +5,23 @@ set -e
 input_vcf=$1
 output_vcf=$2
 
-decompressed_input=${input_vcf//'.gz'/}
-decompressed_output=${output_vcf//'.gz'/}
-bgzip -d $input_vcf -o $decompressed_input
+decompressed_input=$input_vcf
+decompressed_output=$output_vcf
+
+if [[ $input_vcf == *.gz ]]; then
+    decompressed_input=${input_vcf//'.gz'/}
+    bgzip -d $input_vcf -o $decompressed_input
+fi
+if [[ $output_vcf == *.gz ]]; then
+    decompressed_output=${output_vcf//'.gz'/}
+fi
 
 perl -pe '/^((?!^chr).)*$/ && s/^([^#])/chr$1/gsi' \
     "$decompressed_input" > "$decompressed_output"
 
-rm $decompressed_input
-bgzip $decompressed_output
+if [[ $input_vcf == *.gz ]]; then
+    rm $decompressed_input
+fi
+if [[ $output_vcf == *.gz ]]; then
+    bgzip $decompressed_output
+fi
